pyproject.toml
[tool.poetry]
name = "uriagedatadownload"
version = "0.1.0"
description = ""
authors = ["fukui <fukui@beingcorp.co.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.26.1"
load-dotenv = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
